Adds next and prev return values
While it is fairly easy to implement this functionality outside the library it is convenient to provide it with the other return values so it encapsulates everything needed on the pagination template on a single object. Although most template engines let you do the math on the template side there are some that don't or require writing extra helpers.